on commence à bien s'amuser

+++ src/etats/EtatMort.cpp
@@ -22,7 +22,7 @@ EtatMort::EtatMort(App* tApp) : Etat(tApp) {
 
 	int score = manager->getRessources()->getScore();
 	text2.setFont(font);
-	text2.setString("Votre score est de : " + to_string(score));
+	text2.setString("Votre score est de : " + to_string((long double)score));
 	text2.setCharacterSize(32);
 	text2.setColor(sf::Color::White);
 	text2.setStyle(sf::Text::Bold);
